+++ packages/extension-base/src/services/chain-service/handler/EvmChainHandler.ts
@@ -144,7 +144,6 @@ export class EvmChainHandler extends AbstractChainHandler {
 
     try {
       if (tokenType === _AssetType.ERC721) {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument,@typescript-eslint/no-unsafe-member-access
         tokenContract = new evmApi.api.eth.Contract(_ERC721_ABI, contractAddress);
 
         const [_name, _symbol] = await Promise.all([
@@ -157,7 +156,6 @@ export class EvmChainHandler extends AbstractChainHandler {
         name = _name;
         symbol = _symbol;
       } else {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-argument,@typescript-eslint/no-unsafe-member-access
         tokenContract = new evmApi.api.eth.Contract(_ERC20_ABI, contractAddress);
 
         const [_decimals, _symbol, _name] = await Promise.all([