+++ packages/extension-base/src/services/earning-service/service.ts
@@ -17,7 +17,7 @@ import { addLazy, categoryAddresses, createPromiseHandler, PromiseHandler, remov
 import { fetchStaticCache } from '@subwallet/extension-base/utils/fetchStaticCache';
 import { BehaviorSubject } from 'rxjs';
 
-import { AcalaLiquidStakingPoolHandler, AmplitudeNativeStakingPoolHandler, AstarNativeStakingPoolHandler, BasePoolHandler, BifrostLiquidStakingPoolHandler, BifrostMantaLiquidStakingPoolHandler, InterlayLendingPoolHandler, NominationPoolHandler, ParallelLiquidStakingPoolHandler, ParaNativeStakingPoolHandler, RelayNativeStakingPoolHandler, StellaSwapLiquidStakingPoolHandler } from './handlers';
+import { AcalaLiquidStakingPoolHandler, AmplitudeNativeStakingPoolHandler, AstarNativeStakingPoolHandler, BasePoolHandler, BifrostLiquidStakingPoolHandler, BifrostMantaLiquidStakingPoolHandler, InterlayLendingPoolHandler, NominationPoolHandler, ParallelLiquidStakingPoolHandler, ParaNativeStakingPoolHandler, RelayNativeStakingPoolHandler, StellaSwapLiquidStakingPoolHandler, TaoNativeStakingPoolHandler } from './handlers';
 
 const fetchPoolsData = async () => {
   const fetchData = await fetchStaticCache<{data: Record<string, YieldPoolInfo>}>('earning/yield-pools.json', { data: {} });
@@ -80,6 +80,10 @@ export default class EarningService implements StoppableServiceInterface, Persis
 
       if (_STAKING_CHAIN_GROUP.amplitude.includes(chain)) {
         handlers.push(new AmplitudeNativeStakingPoolHandler(this.state, chain));
+      }
+
+      if (_STAKING_CHAIN_GROUP.bittensor.includes(chain)) {
+        handlers.push(new TaoNativeStakingPoolHandler(this.state, chain));
       }
 
       if (_STAKING_CHAIN_GROUP.nominationPool.includes(chain)) {
@@ -649,6 +653,7 @@ export default class EarningService implements StoppableServiceInterface, Persis
       this.earningRewardSubject.next(stakingRewardState);
 
       this.earningsRewardQueue = [];
+      this.earningRewardReady.resolve();
     });
   }
 
@@ -695,6 +700,11 @@ export default class EarningService implements StoppableServiceInterface, Persis
   }
 
   earningsRewardInterval: NodeJS.Timer | undefined;
+  earningRewardReady: PromiseHandler<void> = createPromiseHandler<void>();
+
+  waitEarningRewardReady () {
+    return this.earningRewardReady.promise;
+  }
 
   runSubscribeStakingRewardInterval () {
     const addresses = this.state.keyringService.context.getDecodedAddresses();