+++ packages/extension-base/src/services/earning-service/utils/index.ts
@@ -126,6 +126,8 @@ export function isActionFromValidator (stakingType: YieldPoolType, chain: string
     return true;
   } else if (_STAKING_CHAIN_GROUP.para.includes(chain)) {
     return true;
+  } else if (_STAKING_CHAIN_GROUP.bittensor.includes(chain)) {
+    return true;
   }
 
   return false;