+++ packages/extension-koni-ui/src/components/Layout/base/Home.tsx
@@ -3,31 +3,52 @@
 
 import { Layout } from '@subwallet/extension-koni-ui/components';
 import { CUSTOMIZE_MODAL } from '@subwallet/extension-koni-ui/constants/modal';
+import { useSelector } from '@subwallet/extension-koni-ui/hooks';
+import { RootState } from '@subwallet/extension-koni-ui/stores';
+import { ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { ButtonProps, Icon, ModalContext } from '@subwallet/react-ui';
-import { FadersHorizontal, MagnifyingGlass } from 'phosphor-react';
+import CN from 'classnames';
+import { BellSimpleRinging, FadersHorizontal, MagnifyingGlass } from 'phosphor-react';
 import React, { useCallback, useContext, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
+import styled from 'styled-components';
 
-type Props = {
+interface Props extends ThemeProps {
   children?: React.ReactNode;
   showFilterIcon?: boolean;
   showSearchIcon?: boolean;
+  showNotificationIcon?: boolean;
   onClickFilterIcon?: () => void;
   onClickSearchIcon?: () => void;
   showTabBar?: boolean;
   isDisableHeader?: boolean;
 
-};
+}
 
-const Home = ({ children, isDisableHeader, onClickFilterIcon, onClickSearchIcon, showFilterIcon, showSearchIcon, showTabBar }: Props) => {
+const Component = ({ children, className, isDisableHeader, onClickFilterIcon, onClickSearchIcon, showFilterIcon, showNotificationIcon, showSearchIcon, showTabBar }: Props) => {
   const navigate = useNavigate();
   const { t } = useTranslation();
   const { activeModal } = useContext(ModalContext);
+  const { unreadNotificationCountMap } = useSelector((state: RootState) => state.notification);
+  const { currentAccountProxy, isAllAccount } = useSelector((state: RootState) => state.accountState);
+  const { notificationSetup: { isEnabled: notiEnable } } = useSelector((state: RootState) => state.settings);
+
+  const unreadNotificationCount = useMemo(() => {
+    if (!currentAccountProxy || !unreadNotificationCountMap) {
+      return 0;
+    }
+
+    return isAllAccount ? Object.values(unreadNotificationCountMap).reduce((acc, val) => acc + val, 0) : unreadNotificationCountMap[currentAccountProxy.id] || 0;
+  }, [currentAccountProxy, isAllAccount, unreadNotificationCountMap]);
 
   const onOpenCustomizeModal = useCallback(() => {
     activeModal(CUSTOMIZE_MODAL);
   }, [activeModal]);
+
+  const onOpenNotification = useCallback(() => {
+    navigate('/settings/notification');
+  }, [navigate]);
 
   const headerIcons = useMemo<ButtonProps[]>(() => {
     const icons: ButtonProps[] = [];
@@ -60,8 +81,26 @@ const Home = ({ children, isDisableHeader, onClickFilterIcon, onClickSearchIcon,
       });
     }
 
+    if (showNotificationIcon) {
+      icons.push({
+        icon: (
+          <div className={'notification-icon'}>
+            <Icon
+              phosphorIcon={BellSimpleRinging}
+              size='md'
+            />
+            {notiEnable && !!unreadNotificationCount && <div className={CN('__unread-count')}>{unreadNotificationCount}</div>}
+          </div>
+
+        ),
+        onClick: onOpenNotification,
+        tooltip: t('Notifications'),
+        tooltipPlacement: 'bottomRight'
+      });
+    }
+
     return icons;
-  }, [onClickFilterIcon, onClickSearchIcon, onOpenCustomizeModal, showFilterIcon, showSearchIcon, t]);
+  }, [onClickFilterIcon, onClickSearchIcon, onOpenCustomizeModal, onOpenNotification, showFilterIcon, showNotificationIcon, showSearchIcon, t, unreadNotificationCount, notiEnable]);
 
   const onClickListIcon = useCallback(() => {
     navigate('/settings/list');
@@ -69,6 +108,7 @@ const Home = ({ children, isDisableHeader, onClickFilterIcon, onClickSearchIcon,
 
   return (
     <Layout.Base
+      className={className}
       headerCenter={false}
       headerIcons={headerIcons}
       headerLeft={'default'}
@@ -83,5 +123,31 @@ const Home = ({ children, isDisableHeader, onClickFilterIcon, onClickSearchIcon,
     </Layout.Base>
   );
 };
+
+const Home = styled(Component)<Props>(({ theme: { token } }: Props) => ({
+  '.ant-sw-header-right-part': {
+    display: 'flex'
+  },
+
+  '.notification-icon': {
+    position: 'relative',
+    display: 'flex'
+  },
+
+  '.__unread-count': {
+    borderRadius: '50%',
+    color: token.colorWhite,
+    fontSize: token.sizeXS,
+    fontWeight: token.bodyFontWeight,
+    lineHeight: token.lineHeightLG,
+    paddingTop: 0,
+    paddingBottom: 0,
+    backgroundColor: token.colorError,
+    position: 'absolute',
+    right: 0,
+    bottom: 0,
+    minWidth: '12px'
+  }
+}));
 
 export { Home };